08_streamlit_app/08_langGraph/requirements.txt
# Core LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# OpenAI
openai>=1.50.0

# Streamlit UI
streamlit>=1.40.0

# Data / utils
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: persistent SQLite checkpointing
# langgraph-checkpoint-sqlite

# Optional: vector store for taste memory
# chromadb
# sentence-transformers
